DefaultController - перечень форумов. ForumController - перечень топиков. TopicController - перечень постов. PostController - действия с постом (пока что добавление). Внес поправки в заполнитель/fixture, маршрутизацию, тест. - В заполнителе теперь верно указано кол-во топиков в форуме PHP Добавил учет кол-ва постов при их написании.
