+++ src/ForumBundle/Resources/config/routing/topic.yml
@@ -1,6 +1,6 @@
 topic_show:
     path:     /{id}/{page}
-    defaults: { _controller: "ForumBundle:Post:index", page: 1 }
+    defaults: { _controller: "ForumBundle:Topic:index", page: 1 }
     requirements:
         id:   \d+
         page: \d+
